+++ livre/eq/Config.cpp
@@ -39,6 +39,7 @@
 
 #ifdef LIVRE_USE_ZEQ
 #  include <livre/eq/zeq/communicator.h>
+#  include <zerobuf/render/imageJPEG.h>
 #endif
 
 #include <eq/eq.h>
@@ -66,6 +67,7 @@ public:
     Boxf volumeBBox;
 #ifdef LIVRE_USE_ZEQ
     std::unique_ptr< zeq::Communicator > communicator;
+    ::zerobuf::render::ImageJPEG imageJPEG;
 #endif
     bool redraw;
     Vector2ui dataFrameRange;
@@ -90,6 +92,25 @@ FrameData& Config::getFrameData()
 const FrameData& Config::getFrameData() const
 {
     return _impl->framedata;
+}
+
+::zerobuf::render::ImageJPEG& Config::getImageJPEG() const
+{
+#ifdef LIVRE_USE_ZEQ
+    return _impl->imageJPEG;
+#else
+    LBTHROW( std::runtime_error( "ZeroEq missing, impossible to send imageJPEG" ));
+#endif
+}
+
+bool Config::renderJPEG()
+{
+    _impl->imageJPEG.setData( std::vector< uint8_t >( ));
+    getFrameData().getFrameSettings().setGrabFrame( true );
+    frame();
+    while( _impl->imageJPEG.getDataVector().empty( ))
+        handleEvents();
+    return true;
 }
 
 const ApplicationParameters& Config::getApplicationParameters() const
@@ -326,9 +347,8 @@ bool Config::switchToViewCanvas( const eq::uint128_t& viewID )
     return true;
 }
 
-void Config::handleEvents()
+void Config::handleNetworkEvents()
 {
-    eq::Config::handleEvents();
 #ifdef LIVRE_USE_ZEQ
     _impl->communicator->handleEvents();
     _impl->communicator->publishHeartbeat();
@@ -394,7 +414,7 @@ bool Config::handleEvent( eq::EventICommand command )
         const uint8_t* dataPtr =
             reinterpret_cast< const uint8_t* >( command.getRemainingBuffer( dataSize ) );
 
-        _impl->communicator->publishImageJPEG( dataPtr, dataSize );
+        _impl->imageJPEG.setData( dataPtr, dataSize );
         return false;
     }
 #endif