+++ livre/eq/zeq/communicator.cpp
@@ -31,6 +31,8 @@
 #include <livre/lib/configuration/VolumeRendererParameters.h>
 
 #include <zerobuf/render/lookupTable1D.h>
+#include <zerobuf/render/frame.h>
+#include <zerobuf/render/imageJPEG.h>
 #include <zeq/zeq.h>
 #include <zeq/hbp/hbp.h>
 #include <lunchbox/clock.h>
@@ -56,6 +58,12 @@ public:
             return;
 
         _setupRequests();
+        _config.getImageJPEG().setRequestedFunction( [&]
+            { return _config.renderJPEG(); });
+        _frame.setRequestedFunction( [&]
+            { return publishFrame(); });
+        _frame.setUpdatedFunction( [&]
+            { return updateFrame(); });
         _setupSubscriber();
         _setupHTTPServer( argc, argv );
     }
@@ -70,13 +78,12 @@ public:
         const auto& frameSettings = _getFrameData().getFrameSettings();
         const auto& params = _config.getApplicationParameters();
 
-        const ::zeq::Event& frame = ::zeq::hbp::serializeFrame(
-                                        ::zeq::hbp::data::Frame(
-                                            params.frames[0],
-                                            frameSettings.getFrameNumber(),
-                                            params.frames[1],
-                                            params.animation ));
-        return _publisher.publish( frame );
+        _frame.setStart( params.frames[0] );
+        _frame.setCurrent( frameSettings.getFrameNumber( ));
+        _frame.setEnd( params.frames[1] );
+        _frame.setDelta( params.animation );
+
+        return _publisher.publish( _frame );
     }
 
     bool publishCamera()
@@ -94,13 +101,6 @@ public:
         }
     }
 
-    bool publishImageJPEG( const uint8_t* data, const uint64_t size )
-    {
-        const ::zeq::hbp::data::ImageJPEG image( size, data );
-        const auto& event = ::zeq::hbp::serializeImageJPEG( image );
-        return _publisher.publish( event );
-    }
-
     bool onRequest( const ::zeq::Event& event )
     {
         const auto& eventType = ::zeq::vocabulary::deserializeRequest( event );
@@ -110,33 +110,25 @@ public:
         return i->second();
     }
 
-    void onFrame( const ::zeq::Event& event )
+    void updateFrame()
     {
-        const auto& frame = ::zeq::hbp::deserializeFrame( event );
-
         if( _config.getDataFrameCount() == 0 )
             return;
 
         auto& frameSettings = _config.getFrameData().getFrameSettings();
         auto& params = _config.getApplicationParameters();
 
-        if( frame.current == frameSettings.getFrameNumber() &&
-            frame.delta == params.animation &&
-            frame.start == params.frames.x() &&
-            frame.end == params.frames.y( ))
+        if( _frame.getCurrent() == frameSettings.getFrameNumber() &&
+            _frame.getDelta() == params.animation &&
+            _frame.getStart() == params.frames.x() &&
+            _frame.getEnd() == params.frames.y( ))
         {
             return;
         }
 
-        frameSettings.setFrameNumber( frame.current );
-        params.animation = frame.delta;
-        params.frames = { frame.start, frame.end };
-    }
-
-    bool requestImageJPEG()
-    {
-        _config.getFrameData().getFrameSettings().setGrabFrame( true );
-        return true;
+        frameSettings.setFrameNumber( _frame.getCurrent() );
+        params.animation = _frame.getDelta();
+        params.frames = { _frame.getStart(), _frame.getEnd() };
     }
 
     bool requestExit()
@@ -161,13 +153,15 @@ private:
 #ifdef ZEQ_USE_HTTPXX
     std::unique_ptr< ::zeq::http::Server > _httpServer;
 #endif
+    ::zerobuf::render::Frame _frame;
     Config& _config;
 
     void _setupRequests()
     {
-        _requests[::zeq::hbp::EVENT_FRAME] = [&]{ return publishFrame(); };
-        _requests[::zeq::hbp::EVENT_IMAGEJPEG] = [&]
-            { return requestImageJPEG(); };
+        _requests[ _frame.getTypeIdentifier() ] = [&]
+            { return publishFrame(); };
+        _requests[ _config.getImageJPEG().getTypeIdentifier( )] = [&]
+            { return _config.renderJPEG(); };
         _requests[::zeq::vocabulary::EVENT_EXIT] = [&]{ return requestExit(); };
         _requests[ _getFrameData().getVRParameters().getTypeIdentifier( )] = [&]
             { return _publisher.publish( _getFrameData().getVRParameters( )); };
@@ -184,12 +178,12 @@ private:
     {
 #ifdef ZEQ_USE_HTTPXX
         _httpServer = ::zeq::http::Server::parse( argc, argv, _subscriber );
+
         if( !_httpServer )
             return;
 
-        // subscriber->registerHandler( ::zeq::hbp::EVENT_FRAME,
-        //                                 std::bind( &Impl::onFrame, this,
-        //                                            std::placeholders::_1 ));
+        _httpServer->add( _frame );
+        _httpServer->add( _config.getImageJPEG() );
         _httpServer->add( _getFrameData().getCameraSettings( ));
         _httpServer->add( _getRenderSettings().getTransferFunction( ));
 #endif
@@ -197,12 +191,10 @@ private:
 
     void _setupSubscriber()
     {
-        _subscriber.registerHandler( ::zeq::hbp::EVENT_FRAME,
-                                     [ this ]( const ::zeq::Event& event )
-                                         { onFrame( event ); } );
         _subscriber.registerHandler( ::zeq::vocabulary::EVENT_REQUEST,
                                      [ this ]( const ::zeq::Event& event )
                                          { onRequest( event ); } );
+        _subscriber.subscribe( _frame );
         _subscriber.subscribe( _getFrameData().getVRParameters( ));
         _subscriber.subscribe( _getFrameData().getCameraSettings( ));
         _subscriber.subscribe( _getRenderSettings().getTransferFunction( ));
@@ -223,11 +215,6 @@ Communicator::Communicator( Config& config, const int argc, char* argv[] )
 
 Communicator::~Communicator()
 {
-}
-
-void Communicator::publishImageJPEG( const uint8_t* data, const uint64_t size )
-{
-    _impl->publishImageJPEG( data, size );
 }
 
 void Communicator::publishHeartbeat()